+++ R/read-png.R
@@ -14,8 +14,8 @@
 #'        Example 2: When reading a grey+alpha image, if \code{rgba = TRUE}, 
 #'        then image will be read into an RGBA array with 4 planes.  Otherwise
 #'        it will be read in as grey+alpha data in an array with 2 planes.
-#' @param flags flags. default: 0. See \code{?spng_decode_flags} for other
-#'        options.  Must be an integer.
+#' @param flags flags. default: 1 (always decode transparency from tRNS chunks). 
+#'        See \code{?spng_decode_flags} for other options.  Must be an integer.
 #' @param avoid_transpose default: FALSE.
 #'
 #' @return R image object of the specified type
@@ -23,7 +23,7 @@
 #' @export
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 read_png <- function(src, type = c('array', 'raster', 'native_raster', 'indexed'), 
-                     rgba = FALSE, flags = 0L, 
+                     rgba = FALSE, flags = 1L, 
                      avoid_transpose = FALSE) {
   .Call(read_png_, src, match.arg(type), rgba, flags, avoid_transpose)
 }